+++ configs/test-utils/src/index.ts
@@ -3,6 +3,6 @@ export { CustomNotError } from './CustomNotError'
 export { Suspend } from './Suspend'
 export { ThrowError, ThrowNull } from './ThrowError'
 export { sleep } from './sleep'
-export const TEXT = 'TEXT' as const
-export const ERROR_MESSAGE = 'ERROR_MESSAGE' as const
-export const FALLBACK = 'FALLBACK' as const
+export const TEXT = 'TEXT'
+export const ERROR_MESSAGE = 'ERROR_MESSAGE'
+export const FALLBACK = 'FALLBACK'
