+++ examples/next-streaming-react-query/package.json
@@ -11,21 +11,21 @@
   },
   "dependencies": {
     "@suspensive/react": "workspace:*",
-    "@tanstack/react-query": "^5.28.14",
-    "@tanstack/react-query-devtools": "^5.28.14",
-    "@tanstack/react-query-next-experimental": "^5.28.14",
+    "@tanstack/react-query": "^5.36.0",
+    "@tanstack/react-query-devtools": "^5.36.0",
+    "@tanstack/react-query-next-experimental": "^5.36.0",
     "axios": "^1.6.8",
-    "next": "^14.1.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "next": "^14.2.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "tailwindcss": "^3.4.3"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "^14.1.3",
+    "@next/eslint-plugin-next": "^14.2.3",
     "@suspensive/eslint-config": "workspace:*",
     "@suspensive/tsconfig": "workspace:*",
-    "@types/react": "^18.2.75",
-    "@types/react-dom": "^18.2.24",
+    "@types/react": "^18.3.2",
+    "@types/react-dom": "^18.3.0",
     "autoprefixer": "^10.4.19",
     "postcss": "^8.4.38"
   }
