+++ examples/next-streaming-react-query/src/react-query/index.ts
@@ -1,6 +1,1 @@
-import { QueryClient } from '@tanstack/react-query'
-import { cache } from 'react'
-
-export const getQueryClient = cache(() => new QueryClient())
-
 export { SuspenseQuery } from './SuspenseQuery'
