+++ package.json
@@ -43,32 +43,32 @@
     "@arethetypeswrong/cli": "^0.15.3",
     "@changesets/changelog-github": "^0.5.0",
     "@changesets/cli": "^2.27.1",
-    "@commitlint/cli": "^19.2.1",
-    "@commitlint/config-conventional": "^19.1.0",
+    "@commitlint/cli": "^19.3.0",
+    "@commitlint/config-conventional": "^19.2.2",
     "@testing-library/dom": "^9.3.4",
-    "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^14.3.0",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^14.3.1",
     "@testing-library/user-event": "^14.5.2",
-    "@types/node": "^18.19.31",
-    "@vitest/browser": "^1.4.0",
-    "@vitest/coverage-istanbul": "^1.4.0",
-    "@vitest/ui": "^1.4.0",
+    "@types/node": "^18.19.33",
+    "@vitest/browser": "^1.6.0",
+    "@vitest/coverage-istanbul": "^1.6.0",
+    "@vitest/ui": "^1.6.0",
     "eslint": "^8.57.0",
     "husky": "^9.0.11",
     "jsdom": "^24.0.0",
     "lint-staged": "^15.2.2",
     "ms": "3.0.0-canary.1",
     "packlint": "^0.2.4",
-    "playwright": "^1.43.0",
+    "playwright": "^1.44.0",
     "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.13",
+    "prettier-plugin-tailwindcss": "^0.5.14",
     "publint": "^0.2.7",
-    "rimraf": "^5.0.5",
-    "sherif": "^0.8.1",
+    "rimraf": "^5.0.7",
+    "sherif": "^0.8.4",
     "tsup": "^8.0.2",
     "turbo": "latest",
-    "typescript": "^5.4.4",
-    "vite": "^5.2.8",
-    "vitest": "^1.4.0"
+    "typescript": "^5.4.5",
+    "vite": "^5.2.11",
+    "vitest": "^1.6.0"
   }
 }
