+++ packages/react-await/package.json
@@ -57,7 +57,7 @@
     "test:watch": "vitest --ui --coverage --typecheck"
   },
   "dependencies": {
-    "use-sync-external-store": "^1.2.0"
+    "use-sync-external-store": "^1.2.2"
   },
   "devDependencies": {
     "@suspensive/eslint-config": "workspace:*",
@@ -65,11 +65,11 @@
     "@suspensive/test-utils": "workspace:*",
     "@suspensive/tsconfig": "workspace:*",
     "@suspensive/tsup": "workspace:*",
-    "@types/react": "^18.2.75",
-    "@types/react-dom": "^18.2.24",
+    "@types/react": "^18.3.2",
+    "@types/react-dom": "^18.3.0",
     "@types/use-sync-external-store": "^0.0.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "peerDependencies": {
     "react": "^16.8 || ^17 || ^18"
