+++ packages/react-await/src/Await.tsx
@@ -119,7 +119,7 @@ class AwaitClient {
         .then((data) => {
           newAwaitState.data = data
         })
-        .catch((error) => {
+        .catch((error: unknown) => {
           newAwaitState.error = error
         }),
     }
